+++ packages/types/CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Changed
+- Use ds processor types from `@subql/types-core` (#2357)
 
 ## [3.4.0] - 2024-03-28
 ### Added